refactor(ruby/examples): single examples/sinatra/ using PayKit (Ludo r3306844575, r3306849327)
Per Ludo's two reviews: drop both legacy MPP-only examples and keep a
single PayKit-driven example at the canonical 'examples/sinatra/' path.

- Delete examples/simple-server/ (bare WEBrick driving Mpp::Server::Charge).
- Delete examples/sinatra/{app,server,config}.rb (the Mpp::Sinatra helper demo).
- Move examples/pay-kit-sinatra/* to examples/sinatra/ (app.rb, config.ru,
  pay_kit.rb, README.md).
- ExampleTest now boots PayKitSinatraExample with PAY_KIT_* env.
- README: 'Examples' section becomes 'Example', repo layout collapsed.

259 tests, 0 failures. Line 98.4% / branch 90.39%.
