+++ ruby/examples/sinatra/pay_kit.rb
