+++ ruby/test/example_test.rb
@@ -8,14 +8,12 @@ class ExampleTest < Minitest::Test
 
   def test_sinatra_example_loads_and_exposes_health_route
     with_env(
-      "MPP_FEE_PAYER_SECRET_KEY" => JSON.generate(Array.new(64, 1)),
-      "MPP_MINT" => "So11111111111111111111111111111111111111112",
-      "MPP_PAY_TO" => pubkey(2),
-      "PORT" => "4568"
+      "PAY_KIT_PAY_TO" => pubkey(2),
+      "PAY_KIT_MPP_SECRET" => "test-secret"
     ) do
       require_relative "../examples/sinatra/app"
 
-      response = Rack::MockRequest.new(RubyMppSinatraExample).get("/health")
+      response = Rack::MockRequest.new(PayKitSinatraExample).get("/health")
 
       assert_equal 200, response.status
       assert_equal({"ok" => true}, JSON.parse(response.body))
